backend/recipes/views.py
from .models import Recipe
from rest_framework.viewsets import ModelViewSet
from .serializers import RecipeSerializer


class RecipeViewSet(ModelViewSet):
    queryset = Recipe.objects.all()
    serializer_class = RecipeSerializer
    # pagination_class = PageNumberPagination
